cm-api/requirements.txt
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==38.0.1
Deprecated==1.2.13
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jwcrypto==1.4.2
MarkupSafe==2.1.1
Naked==0.1.31
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.4.0
pymigratedb==1.2.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.28.0
shellescape==3.8.1
SQLAlchemy==1.4.31
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
